Lots of work, probably last commit for version 1.2:
# Goal graphs
When statistics are enabled, you can select any ClickUp item (Space, List, Folder, or Task), for this item you can then set a custom work goal. Adding a goal can be done from the settings page. Select the type, past the id (can be found in the link to the item), and select the goal amount. These goals will then be dynamically added to the statistics panel.

# Small fixes
* Fix some bugs revolving time, and date formats.
* removed logs
* Simplified the basing of events to the statistics pane
* Removed the background of the waze.gif.

# Further work
These are things I did not see worth the time, but I'll document them here in case somebody else wants to take care of them.
* Statistics reloading is not great
* Statistics lay out for smaller screens.
* Dynamic select goes out of bounds when screen gets to small.
* Changing a tracking entry while stats are open breaks everything.
* Updating descriptions seems to be broken, I don't use it but I might end up fixing this in the future.
* styling of statistics and goals panel, if you really want to be fancy.
* Dynamic select upgrade to use the hierarchy select. To make sure you cant make a mistake